+++ links.js
@@ -1,6 +1,8 @@
 const  path = require('path');
 const fs = require('fs');
 const marked = require('marked');
+const { rejects } = require('assert');
+const { resolve } = require('dns');
 //const jsdom = require('jsdom');
 //const { JSDOM } = jsdom;
 
@@ -8,19 +10,21 @@ const marked = require('marked');
 
 
 // si existe ruta 
-
 const isFile = (rutaAbs) => {
     const fileExist = fs.statSync(rutaAbs).isFile();
     return fileExist;
 };
 
   
+
+
 //abrir directorio
 
   const readDirectory=(ruta)=>{ 
       const readDir=fs.readdirSync(ruta,'utf8');
     return readDir;
    } 
+
 
 
 //convertir un archivo md a un contenido  html   
@@ -30,34 +34,41 @@ const mdToHTML = (contenido) => {
   };
 
 
- //obtener los links de los archivos 
+
+ //obtener los links de los archivos dentro de un directorio
+ 
+const filesLinkMd =(rutaAbs)=> {
+       let arrayPaths =[];
+       let filesDirectory=readDirectory(rutaAbs);
+       filesDirectory.forEach (fileOrDirectory=> {
+        let pathJoin=path.join(rutaAbs,fileOrDirectory) ;
+
+        if(fs.statSync(pathJoin).isFile() && path.extname(pathJoin)==='.md') {
+            arrayPaths.push(pathJoin)
+           
+        } else if (fs.statSync(pathJoin).isDirectory()){
+        arrayPaths =arrayPaths.concat(filesLinkMd(pathJoin));
+            
+        }         
+    })
+    return  arrayPaths;
+}
+
+
+let mor=filesLinkMd("D:\\PROGRAMACION\\LIM015-md-links\\pruebas\\carpeta1")
+console.log(mor);
+console.log(showDetails(mor));
+
+
+
+//mostrar finalmnete el array 
+
+
+
+// verificar dentro del documento 
+
  
 
- const fileslinkMD = (rutaAbs) => {
-    let mdFilesArray=[];
-    //leer el directorio 
-    let filesDir = fs.readdirSync(rutaAbs);
-
-    filesDir.forEach(ele=> {
-
-        let joinFile=path.join(rutaAbs,ele);
-
-        console.log(joinFile);
-
-        if (fs.statSync(joinFile).isFile() && path.extname(joinFile) === '.md') {
-
-            mdFilesArray.push(joinFile);
-
-        } else if ( fs.statSync(joinFile).isDirectory()){ 
-
-            mdFilesArray=mdFilesArray.concat(fileslinkMD(joinFile));
-
-        }
-    });
-    return mdFilesArray;
-};
-
-console.log(fileslinkMD("D:\\PROGRAMACION\\LIM015-md-links\\pruebas"))
 
 
 
@@ -69,36 +80,26 @@ console.log(fileslinkMD("D:\\PROGRAMACION\\LIM015-md-links\\pruebas"))
 
 
 
-  
 
+// salida de datos mediante promesa .
 
-/*const LinksOut= (rutaAbs)=> new Promise ((resolve,reject)=>{
+const LinksOut= (rutaAbs)=> new Promise ((resolve,reject)=>{
        // si no es un archivo 
     if(!isFile(rutaAbs)){
-        const pathArray= files(rutaAbs);
-        resolve (allLinks(pathArray,rutaAbs));
-        }
+
+        const pathArray= filesLinkMd (rutaAbs);
+                resolve (allLinks(pathArray,rutaAbs));
+
+    }else if (path.extname(rutaAbs) ==='.md') {
+        const PathAbsoluteMD=[rutaAbs];
+        resolve (allLinks(PathAbsoluteMD,rutaAbs));
+    }
 
     })
-*/
-
-
-    
 
 
 
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-//module.exports = LinksOut;
+module.exports = LinksOut;
