+++ pruebas/carpeta2/carpeta2..md
@@ -1,0 +1,71 @@
+# PROYECTO DATA DEL STUDIO GHIBLI
+
+Este PROYECTO es una página web  interactiva que va dirigido a cinefilos y faánitcos del ánime que desean visualizar detalles  de sus peliculas favoritas del estudio Ghibli , que es el estudio japonés de animación, considerado por la crítica especializada y muchos cinéfilos como uno de los mejores estudios de animación del mundo en la actualidad.
+
+Toda esta información puede ser manejada por los usuarios, de tal manera que puedan filtrarla y ordenarla según diversas especificaciones.
+
+## Contenido
+
+* [1. Detalles de la página web](#1-detalles-de-la-página-web)
+
+* [2. Historias de usuario](#2-historias-de-usuario)
+
+* [3. Prototipo de alta fidelidad](#3-prototipo-de-alta-fidelidad)
+
+* [4. Plan de acción](#4-plan-de-acción)
+
+***
+
+## 1. Detalles de la página web
+
+![paginaWeb1.JPG](./src/img/paginaWeb1.JPG)
+
+visita nuestra página en este enlace  [data Gibli pages ](https://ninoska2000.github.io/LIM015-data-lovers-nat/src/) del proyecto.
+
+
+La página Web elaborada consta de secciones de header, donde se muestra las opciones de tres páginas ,de tal manera que el usuario puede visualizar las colecciones del estudio GHibli mediante la opción de Collection , con la opción de characteres puede visualizar todo los personajes del estudio Ghibli y la última opción de Rankings, Puedes visualizar los Rankings de peliculas mas taquilleras del estudio Ghibli .
+
+La página Collection y Characters cuenta dentro de ella con opciones de filtrado, ordenado y/o búsqueda de tal manera que pueda facilitar al usuario su fácil navegación.
+
+
+## 2. Historias de usuario
+
+Una historia de usuario es una explicación general e informal de una función de software escrita desde la perspectiva del usuario final. Su propósito es articular cómo proporcionará una función de software valor al cliente.
+
+En este proyecto abordamos 4 Historias de usuarios:
+
+[![historiaS.JPG](./src/img/historiaS.JPG) ](https://lucid.app/lucidspark/52e68895-0b35-4aff-8231-835cb66cf848/edit?referringApp=slack&shared=true#)
+
+
+
+
+## 3. Prototipo de alta fidelidad
+
+### Aqui se muestran los prototipos:
+
+![Prototipo de figma](./src/img/EstudioGibliPrototipo.jpg)
+
+![Prototipo de figma móbil 1](./src/img/PrototioCelular1.jpg)
+
+![Prototipo de figma móbil 2](./src/img/PrototioCelular2.jpg)
+## 4. Plan de acción
+1. Conocer al usuario.
+2. Armar las historias de usuario y criterios de aceptación.
+3. Elaborar prototipo de alta fidelidad en Figma (Desktop y móbil).
+4. Feedback de usuarios para mejorar prototipo.
+5. Mejorar el prototipo.
+6. Implementar la interfaz (html/css/js).
+8. Sección de películas: Mostrar los posters.
+9. Sección de películas: Colocar modal donde se muestra título, descripción, etc.
+10. Sección de películas:
+    - Filtrado por directores.
+    - Ordenado por año.
+11. Sección de personajes:
+    - Filtrado por tipo de especies.
+    - Ordenado de la A a la Z.
+    - Ordenado de la Z a la A.
+
+Agregar las secciones
+* Sección de personajes: Mostrar los personajes.
+* Sección de personajes: Calcular cantidad de personajes.
+* Botones en el modal: Mostrar todos los personajes, locaciones y vehicles de acuerdo a la pelicula seleccionada.
