+++ pruebas/general.md
@@ -1,0 +1,48 @@
+# SECRET DOCTOR
+SECRET DOCTOR es una aplicación web que resultará útil tanto para el publico en general como para uso del personal médico. Debido a que nuestros usuarios tendran por primera vez la oportunidad de compartir los resultados de los examenes de sus pacientes de manera incognita, al enviar los resultados de los examenes  en lenguaje cifrado ya no correrán el riesgo que se divulgen con personas equivocadas.
+
+
+![alt text](imagenes/resultados.jpg)
+
+### Definición del Producto
+
+* Público Objetivo:
+
+  - Todas las personas que tienen interes en compartir los resultados de sus examenes médicos de  forma secreta, creando una  confidencialidad profesional.
+
+    ![alt text](imagenes/analisisApp.jpg)
+
+* Cuáles son los objetivos de estos usuarios en relación con tu producto.
+
+  - Compartir los analisis solo con quienes deseen mediante un mensaje cifrado sin temor a que pueda ser filtrado, nadie podra recuperar los resultados del analisis sin la clave secreta.
+
+    ![alt text](imagenes/cifrado.jpeg)  
+
+* ¿Cómo crees que el producto que estás creando está resolviendo sus problemas?
+
+  - Protegiendo la confidencialidad de su informacion.
+  - Se sentiran seguros de compartir los resultados de sus examenes mediante un algoritmo de cifrado de textos.
+
+### Instrucciones de uso
+  Para acceder a la aplicación web "a" se tiene que seguir el siguiente link:
+
+  "https://ninoska2000.github.io/LIM015-cipher/"
+
+  La página tiene una interfaz sencilla y de fácil uso con una pantalla para todas las acciones, lo que el usuario tendría que hacer es:
+* Abrir la página
+* En el primer recuadro digitar el resultado de los examenes realizado al paciente  que desea codificar.
+* En el segundo recuadro indicar la clave secreta en numeros.
+* Dar click en el botón respecto a la acción que desea realizar(encriptar o desincriptar)
+* Tendra el resultado deseado en el recuadro final.
+
+### Decisiones de Diseño
+
+la plataforma virtual esta diseñado inicialmente para encriptar textos, pero se ha proyectado para agregar archivos para encriptar y que el usuario final lo pueda descargar.
+
+
+### Consideraciones Técnicas
+* SECRET DOCTOR, pasó todas las pruebas test unitarios
+![alt text](imagenes/test.PNG) 
+
+
+*
